Reorganise and dask implementation
