+++ africanus/averaging/tests/test_time_and_channel_averaging.py
@@ -4,12 +4,17 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from operator import mul
-
 import numpy as np
+from numpy.testing import assert_array_equal
 import pytest
 
-from africanus.compatibility import reduce
+from africanus.averaging.support import unique_time, unique_baselines
+from africanus.averaging.time_and_channel_avg import time_and_channel
+from africanus.averaging.time_and_channel_mapping import (row_mapper,
+                                                          channel_mapper)
+
+nchan = 16
+ncorr = 4
 
 
 @pytest.fixture
@@ -19,116 +24,256 @@ def time():
 
 @pytest.fixture
 def ant1():
-    return np.asarray([0,   0,   1,   0,   0,   1,   2,   0,   0,   1])    # noqa
+    return np.asarray([0,   0,   1,   0,   0,   1,   2,   0,   0,   1],
+                      dtype=np.int32)
 
 
 @pytest.fixture
 def ant2():
-    return np.asarray([1,   2,   2,   0,   1,   2,   3,   0,   1,   2])    # noqa
+    return np.asarray([1,   2,   2,   0,   1,   2,   3,   0,   1,   2],
+                      dtype=np.int32)
+
+
+@pytest.fixture
+def uvw():
+    return np.asarray([[1.0,   1.0,  1.0],
+                       [2.0,   2.0,  2.0],
+                       [3.0,   3.0,  3.0],
+                       [4.0,   4.0,  4.0],
+                       [5.0,   5.0,  5.0],
+                       [6.0,   6.0,  6.0],
+                       [7.0,   7.0,  7.0],
+                       [8.0,   8.0,  8.0],
+                       [9.0,   9.0,  9.0],
+                       [10.0, 10.0, 10.0]])
+
+
+@pytest.fixture
+def interval():
+    data = np.asarray([1.9, 2.0, 2.1, 1.85, 1.95, 2.0, 2.05, 2.1, 2.05, 1.9])
+    return 0.1 * data
+
+
+@pytest.fixture
+def weight(time):
+    shape = (time.shape[0], ncorr)
+    return np.arange(np.product(shape), dtype=np.float64).reshape(shape)
+
+
+@pytest.fixture
+def sigma(time):
+    shape = (time.shape[0], ncorr)
+    return np.arange(np.product(shape), dtype=np.float64).reshape(shape)
+
+
+@pytest.fixture
+def weight_spectrum(time):
+    shape = (time.shape[0], nchan, ncorr)
+    return np.arange(np.product(shape), dtype=np.float64).reshape(shape)
+
+
+@pytest.fixture
+def sigma_spectrum(time):
+    shape = (time.shape[0], nchan, ncorr)
+    return np.arange(np.product(shape), dtype=np.float64).reshape(shape)
 
 
 @pytest.fixture
 def vis():
     def _vis(row, chan, fcorrs):
-        return (np.arange(row*chan*fcorrs, dtype=np.float32) +
-                np.arange(1, row*chan*fcorrs+1, dtype=np.float32)*1j)
+        flat_vis = (np.arange(row*chan*fcorrs, dtype=np.float32) +
+                    np.arange(1, row*chan*fcorrs+1, dtype=np.float32)*1j)
+
+        return flat_vis.reshape(row, chan, fcorrs)
 
     return _vis
 
 
-@pytest.mark.parametrize("corrs", [(1,), (2,), (2, 2)])
-def test_time_and_channel_averaging(time, ant1, ant2, vis, corrs):
-    from africanus.averaging import time_and_channel
+@pytest.fixture
+def flag():
+    def _flag(row, chan, fcorrs):
+        return np.random.randint(0, 2, (row, chan, fcorrs))
 
-    # time = np.asarray([1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 2.0, 3.0, 3.0, 3.0])  # noqa
-    # ant1 = np.asarray([0,   0,   1,   0,   0,   1,   2,   0,   0,   1])    # noqa
-    # ant2 = np.asarray([1,   2,   2,   0,   1,   2,   3,   0,   1,   2])    # noqa
-
-    row = time.shape[0]
-    chan = 5
-    fcorrs = reduce(mul, corrs, 1)
-
-    vis = vis(row, chan, fcorrs).reshape((row, chan) + corrs)
-    flags = np.zeros(vis.shape, dtype=np.uint8)
-
-    # Test no averaging case
-    avg_vis, avg_time, avg_ant1, avg_ant2 = time_and_channel(
-                                time, ant1, ant2, vis, flags,
-                                avg_time=None, avg_chan=None,
-                                return_time=True, return_antenna=True)
-
-    np.testing.assert_array_almost_equal(avg_ant1, ant1)
-    np.testing.assert_array_almost_equal(avg_ant2, ant2)
-    np.testing.assert_array_almost_equal(avg_time, time)
-    np.testing.assert_array_almost_equal(avg_vis, vis)
-
-    # Now do some averaging
-    avg_vis, avg_time, avg_ant1, avg_ant2 = time_and_channel(
-                                time, ant1, ant2, vis, flags,
-                                avg_time=2, avg_chan=2,
-                                return_time=True, return_antenna=True)
-
-    np.testing.assert_array_almost_equal(avg_time,
-                                         [1.0, 1.5, 1.5, 2.0, 2.5, 3.0, 3.0])
-
-    np.testing.assert_array_almost_equal(avg_ant1, [0, 0, 1, 2, 0, 0, 1])
-    np.testing.assert_array_almost_equal(avg_ant2, [2, 1, 2, 3, 0, 1, 2])
-
-    # Same correlation shape
-    assert vis.shape[2:] == avg_vis.shape[2:] == corrs
-
-    # This works if we comment out both time and channel
-    # bin normalisation in time_and_channel
-    # assert vis.sum() == avg_vis.sum()
+    return _flag
 
 
-@pytest.mark.parametrize("corrs", [(1,), (2,), (2, 2)])
-def test_dask_time_and_channel_averaging(time, ant1, ant2, vis, corrs):
+def _gen_testing_lookup(time, interval, ant1, ant2, flag_row, time_bin_secs,
+                        row_meta):
     """
-    This doesn't test much and especially doesn't not test that the
-    numpy version exactly matches that produced by the dask version.
-    This is because the dask version does not average across chunk
-    boundaries.
+    Generates the same lookup as row_mapper, but different.
+
+    Returns
+    -------
+    list of (float, (int, int), list of lists)
+        Each tuple in the list corresponds to an output row, and
+        is composed of `(avg_time, (ant1, ant2), binned_input_rows)`
+
     """
+    utime, _, time_inv, _ = unique_time(time)
+    ubl, _, bl_inv, _ = unique_baselines(ant1, ant2)
+    bl_time_lookup = np.full((ubl.shape[0], utime.shape[0]), -1,
+                             dtype=np.int32)
+
+    # See :func:`row_mapper` docs for a discussion of this
+    # but we use it to produce different bins for the
+    # TIME_CENTROID/EXPOSURE and TIME/INTERVAL cases
+    sel = flag_row == row_meta.flag_row[row_meta.map]
+    invalid_rows = set(np.where(~sel)[0])
+
+    # Create the row index
+    row_idx = np.arange(time.size)
+
+    # Assign the row indices
+    bl_time_lookup[bl_inv, time_inv] = row_idx
+
+    # Create the time, baseline, row map
+    time_bl_row_map = []
+
+    for bl, (a1, a2) in enumerate(ubl):
+        bl_row_idx = bl_time_lookup[bl, :]
+
+        bin_map = []
+        current_map = []
+
+        flagged_bin_map = []
+        current_flagged_map = []
+
+        for ri in bl_row_idx:
+            if ri == -1:
+                continue
+
+            half_exp = 0.5 * interval[ri]
+
+            # We're starting a new bin
+            if len(current_map) == 0:
+                bin_low = time[ri] - half_exp
+            # Reached passed the endpoint of the bin, start a new one
+            elif time[ri] + half_exp - bin_low > time_bin_secs:
+                bin_map.append(current_map)
+                flagged_bin_map.append(current_flagged_map)
+                current_map = []
+                current_flagged_map = []
+
+            # add current row to the bin
+            if ri not in invalid_rows:
+                current_map.append(ri)
+
+            current_flagged_map.append(ri)
+
+        # Add any remaining maps
+        if len(current_map) > 0:
+            bin_map.append(current_map)
+            flagged_bin_map.append(current_flagged_map)
+
+        # Produce a (avg_time, bl, rows, rows_including_flag_rows) tuple
+        time_bl_row_map.extend((time[rows].mean(), (a1, a2), rows, frows)
+                               for rows, frows
+                               in zip(bin_map, flagged_bin_map))
+
+    # Sort lookup sorted on averaged times
+    return sorted(time_bl_row_map, key=lambda tup: tup[0])
+
+
+@pytest.mark.parametrize("flagged_rows", [
+    [], [8, 9], [4], [0, 1],
+])
+@pytest.mark.parametrize("time_bin_secs", [1, 2, 3, 4])
+@pytest.mark.parametrize("chan_bin_size", [1, 3, 5])
+def test_averager(time, ant1, ant2, flagged_rows,
+                  uvw, interval, weight, sigma,
+                  vis, flag,
+                  weight_spectrum, sigma_spectrum,
+                  time_bin_secs, chan_bin_size):
+
+    time_centroid = time
+    exposure = interval
+
+    vis = vis(time.shape[0], nchan, ncorr)
+    flag = flag(time.shape[0], nchan, ncorr)
+
+    flag_row = np.zeros(time.shape, dtype=np.uint8)
+    flag_row[flagged_rows] = 1
+
+    row_meta = row_mapper(time, interval, ant1, ant2, flag_row, time_bin_secs)
+    chan_map, chan_bins = channel_mapper(nchan, chan_bin_size)
+
+    time_bl_row_map = _gen_testing_lookup(time_centroid, exposure, ant1, ant2,
+                                          flag_row, time_bin_secs,
+                                          row_meta)
+
+    # Input rows associated with each output row
+    row_idx, flag_row_idx = zip(*[(row, flag_rows) for _, _, row, flag_rows
+                                  in time_bl_row_map])
+
+    # Check that the averaged times from the test and accelerated lookup match
+    assert_array_equal([t for t, _, _, _ in time_bl_row_map],
+                       row_meta.time_centroid)
+
+    avg = time_and_channel(time_centroid, exposure, ant1, ant2,
+                           flag_row=flag_row,
+                           time=time, interval=interval, uvw=uvw,
+                           weight=weight, sigma=sigma,
+                           vis=vis, flag=flag,
+                           weight_spectrum=weight_spectrum,
+                           sigma_spectrum=sigma_spectrum,
+                           time_bin_secs=time_bin_secs,
+                           chan_bin_size=chan_bin_size)
+
+    # Take mean time, but first ant1 and ant2
+    expected_time_centroids = [time[i].mean(axis=0) for i in row_idx]
+    expected_times = [time[i].mean(axis=0) for i in flag_row_idx]
+    expected_ant1 = [ant1[i[0]] for i in row_idx]
+    expected_ant2 = [ant2[i[0]] for i in row_idx]
+
+    # Take mean average, but sum of interval and exposure
+    expected_uvw = [uvw[i].mean(axis=0) for i in row_idx]
+    expected_interval = [interval[i].sum(axis=0) for i in flag_row_idx]
+    expected_exposure = [exposure[i].sum(axis=0) for i in row_idx]
+    expected_weight = [weight[i].mean(axis=0) for i in row_idx]
+    expected_sigma = [sigma[i].mean(axis=0) for i in row_idx]
+
+    assert_array_equal(row_meta.time_centroid, expected_time_centroids)
+    assert_array_equal(row_meta.exposure, expected_exposure)
+    assert_array_equal(avg.antenna1, expected_ant1)
+    assert_array_equal(avg.antenna2, expected_ant2)
+    assert_array_equal(avg.time, expected_times)
+    assert_array_equal(avg.uvw, expected_uvw)
+    assert_array_equal(avg.interval, expected_interval)
+    assert_array_equal(row_meta.exposure, expected_exposure)
+    assert_array_equal(avg.weight, expected_weight)
+    assert_array_equal(avg.sigma, expected_sigma)
+
+    out_rows = row_meta.time_centroid.shape[0]
+    chan_avg_shape = (row_meta.exposure.shape[0], chan_bins, flag.shape[2])
+
+    assert avg.vis.shape == chan_avg_shape
+    assert avg.flag.shape == chan_avg_shape
+    assert avg.weight_spectrum.shape == chan_avg_shape
+    assert avg.sigma_spectrum.shape == chan_avg_shape
+
+
+@pytest.mark.parametrize("flagged_rows", [
+    [], [8, 9], [4], [0, 1],
+])
+@pytest.mark.parametrize("time_bin_secs", [1, 2, 3, 4])
+@pytest.mark.parametrize("chan_bin_size", [1, 3, 5])
+def test_dask_averager(time, ant1, ant2, flagged_rows,
+                       uvw, interval, weight, sigma,
+                       vis, flag,
+                       weight_spectrum, sigma_spectrum,
+                       time_bin_secs, chan_bin_size):
+
     da = pytest.importorskip('dask.array')
 
-    from africanus.averaging.dask import time_and_channel
+    from africanus.averaging.dask import time_and_channel as dask_avg
 
-    dask_time = da.concatenate([time]*3)
-    dask_ant1 = da.concatenate([ant1]*3)
-    dask_ant2 = da.concatenate([ant2]*3)
+    rc = (6, 4)
+    fc = (4, 4, 4, 4)
+    cc = (4,)
 
-    rc = dask_time.chunks[0]
-    fc = (5, 5)
+    time = da.from_array(time, chunks=(rc,))
+    interval = da.from_array(interval, chunks=(rc,))
+    ant1 = da.from_array(ant1, chunks=(rc,))
+    ant2 = da.from_array(ant2, chunks=(rc,))
+    vis = da.from_array(vis(sum(rc), sum(fc), sum(cc)), chunks=(rc, fc, cc))
 
-    row = sum(rc)
-    chan = sum(fc)
-    fcorrs = reduce(mul, corrs, 1)
-    avg_time = 2
-    avg_chan = 2
-
-    vis = vis(row, chan, fcorrs).reshape((row, chan) + corrs)
-    dask_vis = da.from_array(vis, chunks=(rc, fc) + corrs)
-    dask_flags = da.zeros(dask_vis.shape, chunks=(rc, fc) + corrs,
-                          dtype=np.uint8)
-
-    avg_vis = time_and_channel(dask_time, dask_ant1, dask_ant2,
-                               dask_vis, dask_flags,
-                               avg_time=avg_time, avg_chan=avg_chan,
-                               return_time=False, return_antenna=False)
-
-    avg_vis, avg_time, avg_ant1, avg_ant2 = time_and_channel(
-                                dask_time, dask_ant1, dask_ant2,
-                                dask_vis, dask_flags,
-                                avg_time=avg_time, avg_chan=avg_chan,
-                                return_time=True, return_antenna=True)
-
-    expected_chans = sum((c + avg_chan - 1) // avg_chan
-                         for c in dask_vis.chunks[1])
-
-    avg_vis = avg_vis.compute()
-    avg_time = avg_time.compute()
-    avg_ant1 = avg_ant1.compute()
-    avg_ant2 = avg_ant2.compute()
-
-    assert avg_vis.shape[1] == expected_chans
